:bug: studyRelation 내용 수정
