--- src/main/java/com/sejong/sejongpeer/domain/StudyRelations/repository/StudyRelationRepository.java
@@ -1,7 +1,0 @@
-package com.sejong.sejongpeer.domain.StudyRelations.repository;
-
-import com.sejong.sejongpeer.domain.StudyRelations.domain.StudyRelation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudyRelationRepository
-        extends JpaRepository<StudyRelation, String>, StudyRelationRepositoryCustom {}
