--- src/main/java/com/sejong/sejongpeer/domain/StudyRelations/repository/StudyRelationRepositoryCustom.java
@@ -1,3 +1,0 @@
-package com.sejong.sejongpeer.domain.StudyRelations.repository;
-
-public interface StudyRelationRepositoryCustom {}
