+++ src/main/java/com/sejong/sejongpeer/domain/studyRelations/repository/StudyRelationRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.sejong.sejongpeer.domain.studyRelations.repository;
+
+public interface StudyRelationRepositoryCustom {
+}
